Move js to separate files
