+++ app/javascript/src/score_modal_trigger.js
@@ -1,0 +1,6 @@
+document.addEventListener("turbolinks:load", function() {
+  $('.modal-trigger').on('click', function (e) {
+    let account = $(this).data()["account"]
+    window.scores_modal.$children[0].update_account(account)
+  })
+});
